finish refactor for icon block. relates to #504.
