+++ src/blocks/icon/controls.js
