+++ src/blocks/icon/edit.js
