+++ src/blocks/icon/inspector.js
@@ -10,7 +10,7 @@ import map from 'lodash/map';
  */
 import applyWithColors from './colors';
 import svg from './svgs';
-import { DEFAULT_ICON_SIZE } from '.././';
+import { DEFAULT_ICON_SIZE } from '.';
 import { MIN_ICON_SIZE, MAX_ICON_SIZE } from './edit';
 
 /**
