+++ src/blocks/icon/svgs.js
